entry: Ensure a local (per transaction) Schema can be used and preferred
Long story short moss is "weird" and has complete filesystem transaction trees,
and invariably we'll see a transaction-local `/usr/lib/os-release` or `/usr/lib/os-info.json`
file that should be preferred to the "tip" system-wide schema.

With this change we respect the branding details in each independent filesystem
state, rather than using the tip file to control the layout of all individual
entries.

Signed-off-by: Ikey Doherty <ikey@aerynos.com>